refactor: OngoingTravel MapCell 에서 OMWMapView로 변경
